add  flag

+++ music_sorter/scripts/sort_music.py
@@ -52,7 +52,9 @@ TAG_OVERRIDE_DICT = {
 @click.option('--ref-artists-file', default=None,
               type=click.Path(exists=True), show_default=True,
               help="Path to a file that contains reference artists (not required).")
-def cli(in_dir, out_dir, keep_album, keep_data, filt, failure_dirname, ref_artists):
+@click.option('--dry-run', is_flag=True, show_default=True,
+              help="Perform a dry run, where the file tags will be processed but music will not be backed up.")
+def cli(in_dir, out_dir, keep_album, keep_data, filt, failure_dirname, ref_artists_file, dry_run):
     # Make sure paths to files the user has specified become absolute and are no longer relative
     # Directory to scan
     in_dir = os.path.abspath(in_dir)
@@ -81,8 +83,8 @@ def cli(in_dir, out_dir, keep_album, keep_data, filt, failure_dirname, ref_artis
     if ok_to_continue:
         # Open a list of artist names we can check against (making the assumption
         # that these names are properly capitalised, etc.).
-        if ref_artists:
-            reference_artists = [x.lower() for x in pd.read_csv("Please pass a path to a file containing names")['name'].values]
+        if ref_artists_file:
+            reference_artists = [x.lower() for x in pd.read_csv(ref_artists_file)['name'].values]
         else:
             reference_artists = None
 
@@ -115,19 +117,21 @@ def cli(in_dir, out_dir, keep_album, keep_data, filt, failure_dirname, ref_artis
         print("\nTop 10 artists by track count that were found:\n{}".format(df['approved_name'].value_counts().head(10)))
         print("The following filetypes were parsed:\n{}\n".format(df['filetype'].value_counts()))
 
-        # Map out the directory structure that will be followed later, then backup to the new destination
-        print("Backing up music that passed checks...")
+        # Map out the directory structure that will be followed later
         df = map_out_directory_structure(df, out_dir, keep_album)
-        backup_files_to_new_directory_structure(df)
-
-        # Apply the same process to failures
-        print("Backing up failures separately...")
         fails = map_out_directory_structure(fails, out_dir, keep_album)
-        backup_files_to_new_directory_structure(fails)
 
-        if keep_data:
-            # Save information to file
-            df.to_csv(os.path.join(out_dir,'musicFileRecord.csv'))
-            fails.to_csv(os.path.join(out_dir,'musicFileRecordFailures.csv'))
+        if not dry_run:
+            # Back-up files
+            print("Backing up music that passed checks...")
+            backup_files_to_new_directory_structure(df)
+            print("Backing up failures separately...")
+            backup_files_to_new_directory_structure(fails)
+
+            if keep_data:
+                # Save information to file
+                df.to_csv(os.path.join(out_dir,'musicFileRecord.csv'))
+                fails.to_csv(os.path.join(out_dir,'musicFileRecordFailures.csv'))
+
 
     return
